+++ core/src/main/kotlin/su/plo/voice/discs/utils/extend/KoinComponent.kt
@@ -1,0 +1,17 @@
+package su.plo.voice.discs.utils.extend
+
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.get
+import kotlin.reflect.KProperty
+
+interface Getter<out T> {
+    val value: T
+}
+
+inline operator fun <T> Getter<T>.getValue(thisRef: Any?, property: KProperty<*>): T = value
+
+inline fun <reified T : Any> KoinComponent.getter(): Getter<T> =
+    object : Getter<T> {
+        override val value: T
+            get() = get()
+    }